% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravel2012.R
\docType{data}
\name{ravel2012}
\alias{ravel2012}
\title{ravel2012}
\usage{
library(ravel2012)
  data(ravel2012)
  print(ravel2012)
}
\value{
Data set as a \code{\link{phyloseq-class}} object
}
\description{
Ravel 2012 vaginal microbiota data set.
}
\details{
This data set contains high-throughput profiling data of the vaginal microbiota collected from reproductive-age women (12-45 years). Exclusion criteria included pregnancy, irregular menstruations, sexual activity, and antibiotic or antimycotic compounds in the past 30 days. See the original publication by Ravel et al. (2012) for further details. We have organized the original public data set into a phyloseq object and prepared the R data package to facilitate the reuse of this data resource. In particular, we expect that it will be useful for scientists who develop methods to model the dynamics of microbial community profiling data.
}
\references{
Ravel J, Gajer P, Abdo Z, Schneider GM, Koenig SS, McCulle SL, et al. Vaginal microbiome of reproductive-age women. Proc Natl Acad Sci U S A (2011) 108(Suppl 1):4680–7. \url{https://doi.org/10.1073/pnas.1002611107}
}
\author{
Binu Mathew \email{binmat@utu.fi}
}
\keyword{data}
