% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravel2012.R
\docType{data}
\name{ravel_2012}
\alias{ravel_2012}
\title{ravel_2012}
\usage{
ravel2012<-ravel2012
}
\value{
Data set as a \code{\link{phyloseq-class}} object
}
\description{
Consists high-throughput profiling data collected from vaginal microbiota samples collected from women, who were not pregnant
, of reproductive age, ranging from 12 to 45 years, regularly menstruating, with a history of sexual activity, and had not taken any antibiotic or antimycotic compounds in the past 30 days.
}
\references{
Ravel J, Gajer P, Abdo Z, Schneider GM, Koenig SS, McCulle SL, et al. Vaginal microbiome of reproductive-age women. Proc Natl Acad Sci U S A (2011) 108(Suppl 1):4680–7. doi:10.1073/pnas.1002611107
}
\author{
Binu Mathew \email{binmat@utu.fi}
}
\keyword{data}
